requirements.txt
azure-functions
playwright==1.44.0
msal
requests
python-dotenv